+++ gekitchen/erd/types/__init__.py
@@ -1,0 +1,4 @@
+"""GE ERD Types"""
+
+from .fridge import *
+from .oven import *
