+++ gekitchen/erd/types/fridge.py
@@ -1,25 +1,19 @@
-"""Types for structured ERD values"""
+"""Types for structured ERD values - fridge"""
 
 __all__ = (
-    'AvailableCookMode',
     'FridgeDoorStatus',
     'FridgeIceBucketStatus',
     'FridgeSetPointLimits',
     'FridgeSetPoints',
     'HotWaterStatus',
     'IceMakerControlStatus',
-    'OvenCookMode',
-    'OvenCookSetting',
-    'OvenConfiguration',
 )
 
 from datetime import timedelta
 from typing import NamedTuple, Optional, TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from .erd_constants import (
-        ErdOvenCookMode,
-        ErdOvenState,
+    from gekitchen.erd.values import (
         ErdDoorStatus,
         ErdFullNotFull,
         ErdOnOff,
@@ -28,41 +22,6 @@ if TYPE_CHECKING:
         ErdPresent,
     )
 
-
-class AvailableCookMode(NamedTuple):
-    """Parsing helper for Available Cook Modes"""
-    byte: int
-    mask: int
-    cook_mode: "ErdOvenCookMode"
-
-
-class OvenConfiguration(NamedTuple):
-    """Cleaner representation of ErdOvenConfiguration"""
-    has_knob: bool
-    has_warming_drawer: bool
-    has_light_bar: bool
-    has_lower_oven: bool
-    has_lower_oven_kitchen_timer: bool
-    raw_value: Optional[str] = None
-
-
-class OvenCookMode(NamedTuple):
-    """Named tuple to represent ErdOvenCookMode for easier formatting later"""
-    oven_state: "ErdOvenState"
-    delayed: bool = False
-    timed: bool = False
-    probe: bool = False
-    warm: bool = False
-    sabbath: bool = False
-
-
-class OvenCookSetting(NamedTuple):
-    """Cleaner representation of ErdOvenCookMode"""
-    cook_mode: OvenCookMode
-    temperature: int
-    raw_bytes: Optional[bytes] = None
-
-
 class FridgeIceBucketStatus(NamedTuple):
     state_full_fridge: "ErdFullNotFull"
     state_full_freezer: "ErdFullNotFull"
@@ -70,11 +29,9 @@ class FridgeIceBucketStatus(NamedTuple):
     is_present_freezer: bool
     total_status: "ErdFullNotFull"
 
-
 class IceMakerControlStatus(NamedTuple):
     status_fridge: "ErdOnOff"
     status_freezer: "ErdOnOff"
-
 
 class FridgeDoorStatus(NamedTuple):
     fridge_right: "ErdDoorStatus"
@@ -83,18 +40,15 @@ class FridgeDoorStatus(NamedTuple):
     drawer: "ErdDoorStatus"
     status: str
 
-
 class FridgeSetPointLimits(NamedTuple):
     fridge_min: int
     fridge_max: int
     freezer_min: int
     freezer_max: int
 
-
 class FridgeSetPoints(NamedTuple):
     fridge: int
     freezer: int
-
 
 class HotWaterStatus(NamedTuple):
     status: "ErdHotWaterStatus"
