+++ gekitchen/erd/values/dishwasher.py
