--- gekitchen/erd_constants/__init__.py
@@ -1,5 +1,0 @@
-"""GE Constants"""
-
-from .erd_codes import ErdCode
-from .erd_constants import *
-from .erd_dishwasher import *
